UnitConverter has functions that take time as seconds (long long) and return the string for year or day of year; added some "test code" at end of GpsPostProcessor main

+++ diff-gps/src/UnitConverter.h
@@ -1,6 +1,7 @@
 #ifndef UNIT_CONVERTER
 #define UNIT_CONVERTER
 
+#include <sstream>
 #include <vector>
 #include "Position.hpp"
 #include "WGS84Ellipsoid.hpp"
@@ -15,8 +16,8 @@ public:
 	UnitConverter();
 	static vector<long double> geoToCart(long double, long double, long double); // Geodetic WGS84 Ellipsoid (degrees/meters) to ECEF Cartesian (meters)
 	static vector<long double> cartToGeo(long double, long double, long double); // ECEF Cartesian (meters) to Geodetic WGS84 Ellipsoid (degrees/meters)
-	static string getYear(int unixTime);
-	static string getDayOfYear(int unixTime);
+	static string getYear(long long unixTime);
+	static string getDayOfYear(long long unixTime);
 
 };
